requirements.txt
discord.py==2.3.2
numpy==2.0.0
